Create README.md
The "Random Password Generator" project is a web application that allows users to generate secure and random passwords. It is built using HTML, CSS, and JavaScript. The user can input the desired length of the password and click a button to generate a random password. The generated password is then displayed on the screen, ready for the user to copy and use for their desired purposes.